First cleanup pass
