+++ src/phantasm-renderer/argument.hh
@@ -4,7 +4,9 @@
 
 #include <phantasm-hardware-interface/arguments.hh>
 
+#include <phantasm-renderer/common/hashable_storage.hh>
 #include <phantasm-renderer/common/state_info.hh>
+
 #include <phantasm-renderer/resource_types.hh>
 
 namespace pr
