+++ src/phantasm-renderer/common/gpu_epoch_tracker.hh
@@ -4,64 +4,35 @@
 
 #include <clean-core/assert.hh>
 
-#include <phantasm-hardware-interface/Backend.hh>
+#include <phantasm-hardware-interface/fwd.hh>
+#include <phantasm-hardware-interface/types.hh>
 
 #include "circular_buffer.hh"
 
 namespace pr
 {
+// ringbuffer for events, keeps track of GPU progress relative to CPU submissions
+// CPU and GPU "epoch": steadily increasing, GPU always behind CPU
+// based on acquiring events and returning them on commandlist submits
+//
+// internally synchronized
 struct gpu_epoch_tracker
 {
-    void initialize(phi::Backend* backend, size_t max_num_events)
+private:
+    struct event_elem
     {
-        _backend = backend;
-        _event_ring = circular_buffer<event_elem>(max_num_events);
-    }
+        phi::handle::event event;
+        uint64_t acquired_epoch_cpu;
+    };
 
-    void destroy()
-    {
-        auto const lg = std::lock_guard(mMutex);
-        while (!_event_ring.empty())
-        {
-            _backend->free(cc::span{_event_ring.get_tail().event});
-            _event_ring.pop_tail();
-        }
-    }
+public:
+    void initialize(phi::Backend* backend, size_t max_num_events);
 
-    [[nodiscard]] phi::handle::event get_event()
-    {
-        {
-            auto const lg = std::lock_guard(mMutex);
-            if (!_event_ring.empty())
-            {
-                // try clearing the oldest event in the buffer
-                auto const tail = _event_ring.get_tail();
-                if (_backend->clearEvent(tail.event))
-                {
-                    // cleared, this event was reached on GPU, advance gpu epoch
-                    _current_epoch_gpu = tail.acquired_epoch_cpu;
-                    // pop and return
-                    _event_ring.pop_tail();
-                    return tail.event;
-                }
-                else
-                {
-                    // oldest event not yet reached, fall through to new event creation
-                }
-            }
-        }
+    void destroy();
 
-        // create a new event
-        return _backend->createEvent();
-    }
+    [[nodiscard]] phi::handle::event get_event();
 
-    void on_event_submission(phi::handle::event event)
-    {
-        auto const lg = std::lock_guard(mMutex);
-        CC_ASSERT(!_event_ring.full() && "event ring full");
-        ++_current_epoch_cpu;
-        _event_ring.enqueue({event, _current_epoch_cpu});
-    }
+    void on_event_submission(phi::handle::event event);
 
     /// returns the epoch that is current on the CPU
     uint64_t get_current_epoch_cpu() const { return _current_epoch_cpu; }
@@ -70,12 +41,6 @@ struct gpu_epoch_tracker
     uint64_t get_current_epoch_gpu() const { return _current_epoch_gpu; }
 
 private:
-    struct event_elem
-    {
-        phi::handle::event event;
-        uint64_t acquired_epoch_cpu;
-    };
-
     std::mutex mMutex;
     phi::Backend* _backend = nullptr;
     uint64_t _current_epoch_cpu = 0;
