+++ src/phantasm-renderer/common/state_info.hh
@@ -1,7 +1,5 @@
 #pragma once
 
-#include <type_traits>
-
 #include <phantasm-hardware-interface/arguments.hh>
 #include <phantasm-hardware-interface/detail/trivial_capped_vector.hh>
 #include <phantasm-hardware-interface/types.hh>
@@ -25,17 +23,6 @@ struct freeable_cached_obj
     murmur_hash hash;
 };
 
-struct pipeline_state_info
-{
-    phi::pipeline_config graphics_config = {};
-    unsigned vertex_size_bytes = 0;
-    bool has_root_consts = false;
-    phi::detail::trivial_capped_vector<phi::vertex_attribute_info, 8> vertex_attributes;
-    phi::detail::trivial_capped_vector<phi::arg::shader_arg_shape, phi::limits::max_shader_arguments> arg_shapes;
-    phi::detail::trivial_capped_vector<murmur_hash, 5> shader_hashes;
-    phi::arg::framebuffer_config framebuffer_config;
-};
-
 // for economic reasons, SRVs, UAVs and Samplers are limited for cache-access shader views
 struct shader_view_info
 {
@@ -44,5 +31,22 @@ struct shader_view_info
     phi::detail::trivial_capped_vector<uint64_t, 4> srv_guids;
     phi::detail::trivial_capped_vector<uint64_t, 4> uav_guids;
     phi::detail::trivial_capped_vector<phi::sampler_config, 2> samplers;
+};
+
+struct graphics_pass_info_data
+{
+    phi::pipeline_config graphics_config = {};
+    unsigned vertex_size_bytes = 0;
+    bool has_root_consts = false;
+    phi::detail::trivial_capped_vector<phi::vertex_attribute_info, 8> vertex_attributes;
+    phi::detail::trivial_capped_vector<phi::arg::shader_arg_shape, phi::limits::max_shader_arguments> arg_shapes;
+    phi::detail::trivial_capped_vector<murmur_hash, 5> shader_hashes;
+};
+
+struct compute_pass_info_data
+{
+    bool has_root_consts = false;
+    phi::detail::trivial_capped_vector<phi::arg::shader_arg_shape, phi::limits::max_shader_arguments> arg_shapes;
+    murmur_hash shader_hash;
 };
 }
