--- src/phantasm-renderer/pipeline_builder.cc
@@ -1,1 +1,0 @@
-#include "pipeline_builder.hh"
