+++ src/phantasm-renderer/pr.hh
@@ -3,9 +3,14 @@
 // This header includes the most important parts of the phantasm renderer used for typical usage
 
 #include "Context.hh"
+#include "format.hh"
+#include "fwd.hh"
 
 #include "CompiledFrame.hh"
+#include "ComputePass.hh"
 #include "Frame.hh"
-#include "Pass.hh"
+#include "GraphicsPass.hh"
+#include "argument.hh"
+#include "pipeline_info.hh"
 
 #include "default_config.hh"
